+++ src/UCTNode.h
@@ -65,8 +65,8 @@ public:
     int get_visits() const;
     float get_policy() const;
     void set_policy(float policy);
-    float get_eval(int tomove) const;
-    float get_raw_eval(int tomove, int virtual_loss = 0) const;
+    double get_eval(int tomove) const;
+    double get_raw_eval(int tomove, int virtual_loss = 0) const;
     float get_net_eval(int tomove) const;
     void virtual_loss();
     void virtual_loss_undo();
